+++ src/FluentSerializer.Json/Converting/Converters/DefaultDateOnlyConverter.cs
@@ -21,7 +21,7 @@ public sealed class DefaultDateOnlyConverter : SimpleTypeConverter<DateOnly>
 	/// <inheritdoc />
 	protected override DateOnly ConvertToDataType(in string currentValue)
 	{
-		var dateValue = currentValue.Length > 2 && currentValue.StartsWith(JsonCharacterConstants.PropertyWrapCharacter) 
+		var dateValue = currentValue.Length >= 2 && currentValue.StartsWith(JsonCharacterConstants.PropertyWrapCharacter) 
 			? currentValue[1..^1]
 			: currentValue;
 		return DateOnly.Parse(dateValue, CultureInfo.CurrentCulture, DateTimeStyles.AllowWhiteSpaces);