+++ src/FluentSerializer.Json/Changelog.md
@@ -4,6 +4,8 @@
 
 - [#91](https://github.com/Marvin-Brouwer/FluentSerializer/issues/91) Improved overall code quality
 - [#93](https://github.com/Marvin-Brouwer/FluentSerializer/issues/93) Fixed ConfigurationStack overriding
+- [#93](https://github.com/Marvin-Brouwer/FluentSerializer/issues/93) Simple type converters now allow empty strings to pass through
+- [#93](https://github.com/Marvin-Brouwer/FluentSerializer/issues/93) Date and Time structures now fail on empty strings
 
 ## v2.2.0